Add extract useExpanded to own file
