+++ packages/common/src/ui/layout/tables/columns/RowExpand/RowExpand.tsx
@@ -1,42 +1,14 @@
 import { Box } from '@mui/material';
-import React, { useCallback } from 'react';
+import React from 'react';
 import { ColumnAlign, ColumnDefinition } from '..';
-import { TableStore, useTableStore } from '../..';
+import { useExpanded } from '../..';
 import { IconButton } from '../../../..';
-
 import {
   ChevronDownIcon,
   ChevronsDownIcon,
   DomainObject,
+  useTableStore,
 } from '../../../../..';
-
-const useExpanded = (rowId: string) => {
-  const selector = useCallback(
-    (state: TableStore) => {
-      return {
-        rowId,
-        isExpanded: state.rowState[rowId]?.isExpanded,
-        toggleExpanded: () => state.toggleExpanded(rowId),
-        expanded: state.numberExpanded > 0,
-      };
-    },
-    [rowId]
-  );
-
-  const equalityFn = (
-    oldState: ReturnType<typeof selector>,
-    newState: ReturnType<typeof selector>
-  ) =>
-    oldState?.isExpanded === newState?.isExpanded &&
-    oldState.rowId === newState.rowId;
-
-  const { isExpanded, toggleExpanded, expanded } = useTableStore(
-    selector,
-    equalityFn
-  );
-
-  return { isExpanded, toggleExpanded, expanded };
-};
 
 export const getRowExpandColumn = <
   T extends DomainObject
@@ -47,9 +19,10 @@ export const getRowExpandColumn = <
   width: 60,
   Header: () => {
     const { numberExpanded, toggleAllExpanded } = useTableStore();
+
     return (
       <IconButton
-        labelKey="app.admin"
+        labelKey="label.expand-all"
         onClick={toggleAllExpanded}
         icon={
           <Box
@@ -58,7 +31,7 @@ export const getRowExpandColumn = <
                 theme.transitions.create('transform', {
                   duration: theme.transitions.duration.leavingScreen,
                 }),
-              transform: !!numberExpanded ? 'rotate(180deg)' : 'rotate(360deg)',
+              transform: !!numberExpanded ? 'rotate(360deg)' : 'rotate(270deg)',
             }}
           >
             <ChevronsDownIcon />
@@ -68,29 +41,24 @@ export const getRowExpandColumn = <
     );
   },
   Cell: ({ rowData }) => {
-    const { toggleExpanded, expanded } = useExpanded(rowData.id);
+    const { toggleExpanded, isExpanded } = useExpanded(rowData.id);
 
     return (
       <IconButton
-        labelKey="app.admin"
+        labelKey="label.expand"
         onClick={event => {
           event.stopPropagation();
           toggleExpanded();
         }}
         icon={
           <Box
-            // sx={{
-            //   animation: !!expanded
-            //     ? `${spin} 1s ease`
-            //     : `${otherSpin} 1s ease`,
-            // }}
             sx={{
               transition: theme =>
                 theme.transitions.create('transform', {
                   easing: theme.transitions.easing.sharp,
                   duration: theme.transitions.duration.leavingScreen,
                 }),
-              transform: expanded ? 'rotate(180deg)' : 'rotate(360deg)',
+              transform: isExpanded ? 'rotate(360deg)' : 'rotate(270deg)',
             }}
           >
             <ChevronDownIcon />
