+++ packages/common/src/ui/layout/tables/components/DataRow/DataRow.tsx
@@ -1,9 +1,9 @@
-import React, { FC, useCallback } from 'react';
+import React, { FC } from 'react';
 import TableRow from '@mui/material/TableRow';
 import TableCell from '@mui/material/TableCell';
 import { Column } from '../../columns/types';
 import { DomainObject } from '../../../../../types';
-import { TableStore, useTableStore } from '../..';
+import { useExpanded } from '../..';
 import { Collapse } from '@mui/material';
 
 interface DataRowProps<T extends DomainObject> {
@@ -13,30 +13,6 @@ interface DataRowProps<T extends DomainObject> {
   rowKey: string;
   ExpandContent?: FC;
 }
-
-const useExpanded = (rowId: string) => {
-  const selector = useCallback(
-    (state: TableStore) => {
-      return {
-        rowId,
-        isExpanded: state.rowState[rowId]?.isExpanded,
-        toggleExpanded: () => state.toggleExpanded(rowId),
-      };
-    },
-    [rowId]
-  );
-
-  const equalityFn = (
-    oldState: ReturnType<typeof selector>,
-    newState: ReturnType<typeof selector>
-  ) =>
-    oldState?.isExpanded === newState?.isExpanded &&
-    oldState.rowId === newState.rowId;
-
-  const { isExpanded, toggleExpanded } = useTableStore(selector, equalityFn);
-
-  return { isExpanded, toggleExpanded };
-};
 
 export const DataRow = <T extends DomainObject>({
   columns,
