+++ packages/common/src/ui/layout/tables/context/TableContext.ts
@@ -1,3 +1,4 @@
+import { useCallback } from 'react';
 import create, { UseStore } from 'zustand';
 import createContext from 'zustand/context';
 
@@ -159,3 +160,30 @@ export const createTableStore = (): UseStore<TableStore> =>
       });
     },
   }));
+
+interface UseExpandedControl {
+  isExpanded: boolean;
+  toggleExpanded: () => void;
+}
+
+export const useExpanded = (rowId: string): UseExpandedControl => {
+  const selector = useCallback(
+    (state: TableStore) => {
+      return {
+        rowId,
+        isExpanded: state.rowState[rowId]?.isExpanded ?? false,
+        toggleExpanded: () => state.toggleExpanded(rowId),
+      };
+    },
+    [rowId]
+  );
+
+  const equalityFn = (
+    oldState: ReturnType<typeof selector>,
+    newState: ReturnType<typeof selector>
+  ) =>
+    oldState?.isExpanded === newState?.isExpanded &&
+    oldState.rowId === newState.rowId;
+
+  return useTableStore(selector, equalityFn);
+};
